api + list + detail
